Create README.md
ElevenLabs test - I thought it would be interesting to see if it would work to use Python to trim the audio after using some extra words to change the mood of the speech. The idea is that you can add to a phrase some extra text like "... she whispered", or "... he said angrily" and get a different sort of output. If you use two commas it makes more of a gap, then you just need to trim it.